--- packages/base-driver/gulpfile.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const gulp = require('gulp');
-const boilerplate = require('@appium/gulp-plugins').boilerplate.use(gulp);
-
-boilerplate({
-  build: 'appium-base-driver',
-});
