+++ packages/base-driver/test/basedriver/README.md
@@ -1,0 +1,5 @@
+# Driver Test Helpers
+
+This directory contains modules which expose test suites that an external driver can use to test against the base implementation.  They are published in the `@appium/base-driver` package.
+
+Drivers wanting to leverage these suites will want to add `@appium/base-driver` to their `devDependencies`.
