--- packages/base-driver/test/basedriver/driver-specs.js
@@ -1,8 +1,0 @@
-// transpile:mocha
-
-import BaseDriver from '../../lib';
-import baseDriverUnitTests from './driver-tests';
-baseDriverUnitTests(BaseDriver, {
-  platformName: 'iOS',
-  'appium:deviceName': 'Delorean'
-});
