+++ packages/base-driver/test/e2e/basedriver/driver.e2e.spec.js
@@ -1,0 +1,8 @@
+// transpile:mocha
+
+import BaseDriver from '../../../lib';
+import baseDriverE2ETests from '../../basedriver/driver-e2e-tests';
+baseDriverE2ETests(BaseDriver, {
+  platformName: 'iOS',
+  'appium:deviceName': 'Delorean'
+});
