+++ packages/base-driver/test/e2e/fixtures/BadZippedApp.zip
