+++ packages/base-driver/test/e2e/fixtures/FakeAndroidApp.asd
