+++ packages/base-driver/test/e2e/fixtures/FakeIOSApp.app.zip
