+++ packages/base-driver/test/unit/basedriver/driver.spec.js
@@ -1,0 +1,9 @@
+// transpile:mocha
+
+import BaseDriver from '../../../lib';
+import { baseDriverUnitTests } from '../../basedriver';
+
+baseDriverUnitTests(BaseDriver, {
+  platformName: 'iOS',
+  'appium:deviceName': 'Delorean'
+});
